Shorten subcloud pre-check error on patch update
When the subcloud management alarm pre-check fails during patch update,
the exception string is added to the failure message in the strategy
step. This might cause its details field to exceed the database limit.

Remove the exception from the saved message and add an instruction for
where to check for more details. Continue logging the exception and its
stack trace to the orchestrator logs.

Test Plan:

PASS:
- Ensure tox tests are passing
- Verify that strategy step fails when subcloud pre-check call fails

Closes-Bug: 1978882

Signed-off-by: Gabriel Silva Trevisan <gabriel.silvatrevisan@windriver.com>
Change-Id: I2ac4f53488882a59c892342139e410b516a18d6f