Don't enable https on subcloud when ca cert is installed
Currently in a DC system, when a ssl_ca certificate is installed,
an incorrect checking in DC sysinv driver triggers https enabling
on subcloud. This change updated the checking so that the CA
certificate is installed but https remains unchanged on subcloud.

Test Plan:
PASS: In DC system with ssl_ca certificate, disable https and uninstall
      ssl_ca certificate on a subcloud. Wait until the ssl_ca
      certificate is synced to the subcloud again. Observe that https
      remains disabled on the subcloud.

Closes-Bug: 1980769
Signed-off-by: Andy Ning <andy.ning@windriver.com>
Change-Id: I136319c8d93eb39d97818924cb07d0d2a013adec