Make Tox compatible to Debian Python3.9 Environment
A different version of astroid is required for python3.9, pylint,
and pep8. The pylint setup related to pip and upper constraints
is different for python3.9. Updated the requirement packages'
software version(to versions that support python3.9) Updated
'cover' tox.ini target to be python3.9

This flake8 suppresion is added:
  H126: flag use of third party mock

These pylint disable/enable alarm checks are removed due to deprecation:
-Enable
  E1606: execfile-builtin
  E1607: old-ne-operator
  E1608: old-octal-literal
  E1610: non-ascii-bytes-literal
  E1611: invalid-unicode-literal
  W1641: eq-without-hash

-Disable
  W0403: relative-import

These pylint suppressions need to be added
  W0707: raise-missing-from
  W1406: redundant-u-string-prefix
  W1514: unspecified-encoding
  E1101: no-member

The explicit addition of python3 checks that were required
when running pylint in python2 to test python3 are removed
since pylint is now running in python3 and gets those
checks implicitly.

Removed config keys from the pylint.rc that are no longer
valid for the version of pylint executed by python3.9.

The zuul target explicitly specifies python3.9 and the
ubuntu-focal node to ensure it has the correct interpreter.

Depends-On: https://review.opendev.org/c/starlingx/distcloud/+/850543
Story: 2008943
Task: 45692

Signed-off-by: BoYuan Chang <boyuan.chang@windriver.com>
Change-Id: I70280a5fee178c228e138d4ac1c047dae7696c85