Fix python3 issue in SubcloudNotFound exception handling
Due to python3 string handling differences, the existing code causes a
new exception: "TypeError: argument of type 'RemoteError' is not iterable"
The fix is to convert the exception to a string before the compare.

Proper subclassing of the SubcloudNotFound exception is a better
long-term solution, and should be considered when centos/python2 is no
longer a requirement.

Test Plan:
- test with condition leading to the initial exception. The
  "TypeError: argument of type 'RemoteError' is not iterable"
  is avoided
- test in Debian and centos

Story: 2010119
Task: 45798

Change-Id: I868555124406c5f3214f675a7254a438e49201ca
Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>