Debian - Fix dcorch audit out-of-sync
On Debian it's observed that the dcorch audit failed due to
a ForeignKeyViolation DBReferenceError raised when updating or
deleting on tables: 'user' and 'role', because they have foreign
key constraints on 'user_id' and 'role_id' referenced by
'user_option' and 'role_option' tables.
The reason that CentOS doesn't have this issue could be it still
use some old versions of db related modules: 'oslo.db', 'psycopg2',
and 'sqlalchemy' on python2. This fix is an enhancement for the
reference relationships, and it would not impact the dcorch audit
function on CentOS.

Test Plan:
PASSED: Dcorch audit on Debian
PASSED: Dcorch audit on Centos

Story: 2010119
Task: 45903

Signed-off-by: Li Zhu <li.zhu@windriver.com>
Change-Id: Ia898bc736b304fcb113c6b610d0ba661c4b9d73f