Include support to json requests on load-import
Change [1] adds an option --local to import the load locally on the
active controller instead of sending it through the API. When this
option is used, the payload of the request will be a json containing
the iso and sig file paths.

This commit adds support to use the new option when load-import is
used with --os-region-name SystemController to import the load to
the dc-vault.

This commit also improves exception handling during load-import,
since it was always returning a generic "server error 500" message.

[1] https://review.opendev.org/c/starlingx/config/+/851082

Test Plan
PASS: import load using "--local" and "--os-region-name
      SystemController" options and verify load imported on
      /var/www/pages/feed and /opt/dc-vault/loads

Regression
PASS: import load using only "--os-region-name SystemController"
      option and verify load imported on /var/www/pages/feed and
      /opt/dc-vault/loads

Story: 2009303
Task: 45886

Relates-to: https://review.opendev.org/c/starlingx/config/+/851082

Change-Id: Iac8c1bcbe7e2cbb380c4450e1a4004e2862c6136
Signed-off-by: Heitor Matsui <HeitorVieira.Matsui@windriver.com>