Adds missing imports to stx-keystone-api-proxy
This change adds the missing imports for both packages: fmclient and
python-barbicanclient to stx-keystone-api-proxy pip packages list.

On DC systemcontroller systems the keystone-api-proxy chart will be
enabled and it needs the fmclient package to work properly, and
therefore  for stx-openstack to be applied on a DC systemcontroller.

Test Plan:
PASS - Build stx-keystone-api-proxy docker image
PASS - Apply stx-openstack application with new stx-keystone-api-proxy
image

Closes-Bug: 1986859

Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>
Change-Id: Ic14695882ddae78c1ab6bceaa38a715eb7541f78