Add "error_description" field on subclouds table
In order to improve the error reporting on subcloud deployment and
upgrading, this changes adds a new field to the subclouds table called
"error_description", which is being filled with the error details
gathered during these stages.

As a first step, a function was added in:
https://review.opendev.org/c/starlingx/distcloud/+/855528

This review will store the log error from that review in the new
error_description field on the subclouds table.

Test Plan:
PASS: Bring up a DC system using an ISO containing these changes.
      Verify that the error_description field is added in the
      database.
PASS: Upgrade the system controller to the load containing these
      changes. Verify that the error_description field is added
      in the database.
PASS: Deploy subclouds generating playbook errors:
      - wrong local registries.
      - invalid SSL certificate.
      Verify that the error_description field is filled with the
      log error that was gathered during deploy.
PASS: Verify that if the error is shorter than 2048 characters, the
      entire string is stored. Otherwise the first 2048 characters
      are stored.

Story: 2010271
Task: 46293

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: Ia686b631f0e2bd4ebb7c41de2f7e7ecf036dda67