Unable to update subcloud details
Adding a fix to solve the API (dcmanager-api) error that Horizon
and CLI returns when updating the description, location, or group
name on subcloud details by using the following string format:

VBOX workstation: workstation: 127.0.0.1

The API responds successfully if no space is placed after the colon
character. The error arises when the user places a space after the
colon character. This happens when the request parameters are
decoded and the values are passed to the YML safe_load method.

In order to force the string and fix the issue, the value was
enclosed in quotes.

Closes-Bug: 1989619

Test Plan - Horizon and CLI:

 - Edit the subcloud from Horizon or run CLI command
 - Update the description, location or group field by adding the following:

   VBOX workstation: workstation: 127.0.0.1

Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>
Change-Id: Ibe1f51d2166c0dffc2234652f5023f3f29cab468