Add subcloud backup delete operation to dcmanager
Add suport for deleting backups for a subcloud or group of subclouds
using the dcmanager API.

This commit includes chages to the subcloud-backup API and dcmanager to
validate user input, create inventory and overrides file, and invoke the
subcloud delete playbook for one or multiple subclouds. The delete
playbook, delivered in a separate commit to the playbooks repo [1],
will delete the subcloud backup files, either on the central archive,
or locally within the subclouds.

Differently from subcloud backup create, the delete operation is to be
executed synchronously, that is, the command will wait for the operation
to complete to return and display any messages returned, if applicable
[2].

Unit tests will be added in a separate commit. Max parallel operations
will be defined during scale testing.

Test Plan:

PASS:
- Ensure tox tests are passing
- Issue backup delete request for a single subcloud
- Issue backup delete request for a small group of subclouds
- Check API reachability using manual HTTP request
- Check validations on API arguments
- Attempt local backup delete on an offline subcloud
- Attempt local backup delete on an unmanaged subcloud
- Attempt local group delete with both valid and invalid subclouds
- Verify backup delete still succeeds when file does not exist
- Verify generated subcloud inventory file
- Verify generated overrides file for different option values
- Verify generated command for delete playbook invoke
- Verify delete playbook is invoked correctly using a stub
- Verify logs and warning message for failures/skips in group
- Verify error message when all subclouds in group fail operation
- Check backup_status and backup_datetime for successful execution
- Test different thread pool sizes for group operation
- Attempt backup delete on subcloud with invalid deploy state
- Integration testing with dcmanager-client
- Integration testing with delete playbook

[1] Review for backup delete wrapper playbook
https://review.opendev.org/c/starlingx/ansible-playbooks/+/857247

[2] Review for subcloud backup delete CLI
https://review.opendev.org/c/starlingx/distcloud-client/+/856312

Story: 2010116
Task: 46210

Signed-off-by: Gabriel Silva Trevisan <gabriel.silvatrevisan@windriver.com>
Change-Id: Ib6aaa1ff555337e6a202d58a1f6f3ad47b133ce2