Fix misleading subcloud backup create error message
Run the operation dcmanager subcloud-backup when a subcloud
is in pre-deploy state throws a misleading error message:
Operation not allowed while subcloud is unmanaged.
This commit prevents the subcloud-backup create command
from running if the subcloud is in a state other than
'complete'.

Test Plan:
1. Add a subcloud using dcmanager subcloud add command.
2. Try to create a backup using dcmanager subcloud-backup
create command while:
- Subcloud online and managed.
- Run the backup-create command while subcloud is in
various states and ensure the command is only accepted
when deploy_status is in 'complete' state.

Closes-bug: 1991972

Signed-off-by: Hugo Brito <hugo.brito@windriver.com>
Change-Id: I36d298b59e3db904addc555e18781765dba45940