Update validations for subcloud-backup operations
The validations have been changed slightly for the subcloud-backup
create and delete operations.
The create command now fails if there are any management affecting
alarms on the specified subcloud(s).
The delete operation does not require that the subclouds have the
'complete' deploy status.
The passing criteria for the restore operation remain unchanged.

Test Plan:
1 PASS - Validate that the create operation will not start if
there are any management affecting alarms on the specified
subcloud(s).
2 PASS - Validate that, without --local-only option, the delete
operation will run when the subcloud(s) is in any state.
3 PASS - Validate that, with the --local-only option, the delete
operation will run only if the subcloud(s) is online and managed,
regardless of the deploy status.
3 PASS - Validate that the other pre-requisites remain unchanged for
the affected commands.

Story: 2010116
Task: 46874

Signed-off-by: Andre Carneiro <Andre.DexheimerCarneiro@windriver.com>
Change-Id: Ibbf5a0dc75a4abbdaae108ae229d215509f7b1dd