Initial integration of DC with admin network
Because the management network and its parameters are
embedded in many parts of the system, having a separate
admin network makes it much easier to change the parameters
of this network (subnet, gateway, etc) after a subcloud has
been provisioned.

The admin network will take precedence over the existing
management network for communication between the
subcloud and system controller if it is defined.

The management network will still exist on the subcloud, but
will be a private network.

This commit enables a subcloud to be initially added
(bootstrapped) with admin network parameters.  Future
commits will be focused on the enablement of subcloud
rehoming, reinstall, and updating for admin network
enabled subclouds.

Test Plan:

- Bootstrap and install DC subcloud with admin network defined.
  PASS: Ensure the expected admin network values are reflected in
        the output of dcmanager subcloud show.
  PASS: Ensure the openstack admin endpoints on both the subcloud
        and system controller for the affected services use the
        admin subnet of the subcloud
  PASS: Ensure the subcloud can become online and in-sync using
        the admin network.
  PASS: AIO-SX, AIO-DX subclouds
  TODO: Standard configuration

Regression:

- Bootstrap and install DC subcloud with no admin network defined.
  PASS: Ensure the openstack admin endpoints on both the subcloud
        and system controller for the affected services use the
        management subnet of the subcloud (no impact)
  PASS: Ensure the subcloud can become online and in-sync with the
        management network (no impact).

Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/865286

Story: 2010319
Task: 46912

Signed-off-by: Steven Webster <steven.webster@windriver.com>
Co-Authored-By: Sirin Shaikh <sirin.shaikh@windriver.com>
Change-Id: I108447026e8e45ba7e2d392e211758745115f1d4