Include storage location in backup status
Includes the storage location (local or central) in the backup status
field. The 'complete' status has been replaced by two new states:
'complete-local' and 'complete-central'. They indicates wether the
backup files are stored locally in the subcloud or in centralized
storage, respectively. The purpose of this change is to make this
difference clearer to the user once the subcloud or a subcloud group
has been backed up.

Test Plan:
1. PASS - Validate that the backup create operation will result in
complete-local backup status if the --local-only option is enabled.
2. PASS - Validate that the backup create operation will result in
complete-central backup status if the --local-only option is
disabled.
3. PASS - Validate that the backup delete operation is unaffected.
4. PASS - Validate that the backup restore operation is unaffected.

Story: 2010116
Task: 46947

Signed-off-by: Andre Carneiro <Andre.DexheimerCarneiro@windriver.com>
Change-Id: Idbed50fa81b152106f91dbbb79c387d1f13746e9