Fix failing zuul jobs due to tox showconfig
As part of how we nest the tox.ini files in STX,
the environments in the sub directories need to also
be defined in the tox.ini in the base file.

I suspect the tox_extra_args change here may have exposed
the problem in some of our tox files.
https://review.opendev.org/c/zuul/zuul-jobs/+/866935

Partial-Bug: #1997255

Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Ia963897be5f844599d2f6145f472f5eac1df6805