Skip prestage_prepare for upgrade on Debian
Since the ostree repo will no longer be copied to /var/run before
prestaging for upgrade, preparing the prestage packages is not
required anymore.

This commit removes support for preparing prestage packages for
upgrade in dcmanager.

Since the existing unit test for this support is no longer
relevant, it has been removed.

Test Plan:
PASS: Verify that prestage for upgrade is successfully executed.

PASS: Verify that prestage for reinstall is successfully executed.

Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/870682
Closes-Bug: 2003271
Change-Id: Ibc4f1a3969db060f708592dbedd91a1f3fdbd6fa
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>