Clear 280.003 alarm on subcloud deletion
When a subcloud was deleted, the 280.003 alarm  was not cleared.
The change was to add the 280.003 alarm to the list of alarms that are checked.

Test Plan:

PASS: Install and bootstrap DC.
PASS: create a subcloud backup and power off the subcloud after reaching:
"Run subcloud1 backup playbook".
PASS: Wait for the alarm 280.003 to be set, then umanage and
delete the subcloud and check if the 280.003 alarm was cleared.

Closes-bug: 2004638

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: I70458b1fe6da4b9c0bf458ee57720a4ffaeafbfb