Check if there is an partial-applied patch when subcloud add is run
In order to avoid the subcloud, which was added
after a patch has been applied but not installed,
to run the patched software, a validation
was added and now it's not possible to add
a subcloud when there is an partial-applied
patch.

Test Case:

PASS: Install and bootstrap DC.
PASS: Upload and apply a patch.
PASS: Run subcloud add command and check if the
output is "Patching operation is in progress."
and check if no subcloud was added.

Closes-bug: 2006939

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: I4113523b6d27aa1f4cda89a56478a5f0833d6095