Check upgrade health before subcloud upgrade orchestration
One of the pre-checks for the subcloud upgrade orchestration includes
a call to get_system_health for that subcloud. This change replaces it
with get_system_health_upgrade. The get_system_health_upgrade
calls get_system_health internally.
This change lets us do additional checks for the subcloud to be ready
for the upgrade such as to check if active kubernetes version is the
latest kubernetes version and allows to extend this check-list in the
future.

Note: Outputs of commands 'system health-query' and
      'system health-query-upgrade' are almost same when the load
      is not imported. The only difference is the message
      'No imported load found. Unable to test further'.
      This can be automatically ignored as,
      1. All other required health checks are performed except
         imported load check.
      2. The message does not contain the keyword 'Fail'.

DC subcloud orchestrated upgrade testing:
PASS: Install system controller: DX + 1 worker with stx 7.0
      and k8s 1.22.5. Add an AIO-SX subcloud with k8s 1.22.5.
      Start system controller platform upgrade as a part of DC
      subcloud orchestrated upgrade. Upgrade health check fails.
      Upgrade system controller k8s to 1.23.1. Upgrade health check
      passes. Perform system controller platform upgrade to stx 8.0.
      After successful platform upgrade, observe load_sync_status and
      kubernetes_sync_status are out_of_sync.
      Create a dcmanager upgrade-strategy for the only subcloud.
      Apply the strategy. Precheck fails. Delete the strategy.
      Manually upgrade k8s on the subcloud to 1.23.1.
      Recreate the upgrade strategy and apply it.
      Precheck passes this time. Upgrade proceeds the precheck.

Story: 2009932
Task: 44815

Signed-off-by: Kaustubh Dhokte <kaustubh.dhokte@windriver.com>
Change-Id: I49d2fa1301b428e3d71a7a15ea706249ffc95713