Increment timeout of DC K8s Root CA strategy
Recently, the internal pod rollout script used in the stages
"trust-both-cas" and "trust-new-ca" of the Kubernetes Root CA update
process was increased from ~8 minutes to ~30 minutes, so the max
possible timeout for the root CA update increased in ~44 minutes.
This change increases the timeout for subclouds, used by dcmanager
kube-rootca-update-strategy, from 60 minutes to 120 minutes.

Test Plan:

PASS: In a DC deploy with 1 subcloud, successfully apply a dcmanager
kube-rootca-update-strategy on the subcloud.

PASS: Repeat the test above, but artificially make the root CA update
process in the subcloud take more than 1 hour to complete and check
that no timeout occurs on the central cloud.

Closes-Bug: 2004594
Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I6460846eba6a37d1d0cdc634ea4cf1314b1b6bc4