Add static content for error messages
This commit is part of the addition of static content to the
subcloud error reporting.
The content is intended to be consumed by both CLI and GUI clients.

There are definitions for errors related to the failed deployment
status and for specific messages depending on the logged `err_code`.
The dictionary does not have a message for every status but can be
extended in the future.

Story: 2010271
Task: 47504

Test Plan
PASS: Install ISO containing these changes.
      Add different subclouds forcing known errors, e.g.:
      * Bootstrap using wrong license.
      * Bootstrap subcloud with misconfigured docker registries.
      * Install subcloud with wrong BMC IP.
      * Install subcloud with wrong BMC credentials.
PASS: Run `dcmanager subcloud errors <<subcloud_name>>
      Check that the output includes the static content related to the
      deploy-status field.
      Check that the output includes the static content related to the
      err_code from original message.
PASS: Open starlingx dashboard (Horizon), go to the subcloud list and
      expand the details of the subclouds.
      Check that the `Subcloud Errors` field includes the static
      content related to the deploy-status field.
      Check that the `Subcloud Errors` field includes the static
      content related to the err_code from original message.
PASS: Force a failure whose err_code is not defined in the dictionary,
      (e.g. Errcode=unrec_pre_unlock, by the time of this review).
      Check that no extra content is presented in both clients.

Signed-off-by: Agustin Carranza <agustin.carranza@windriver.com>
Change-Id: Ic3befd05f118016647ade2f366ba99953520253e