Add release optionality to subcloud-deploy upload/show
Add an optional --release parameter to subcloud-deploy upload/show
commands to enable release optionality in subcloud-deploy upload and
show.

Test Plan:
PASS: Verify that the deployment files were successfully uploaded to
the directory dedicated for that specified release.
PASS: Verify that the deployment files were successfully uploaded to
the directory dedicated for the active release when the '--release'
parameter not present.
PASS: Verify the deployment files that were showed belong to
the specified release.
PASS: Verify the deployment files that were showed belong to
the active release when the '--release' not present.
PASS: Verify that 'None' files were shown when an unknown release
was given to the "subcloud-deploy show" command.

Story: 2010611
Task: 47663

Signed-off-by: lzhu1 <li.zhu@windriver.com>
Change-Id: I402d95a0fd6c5b7818539d2756fc0c13cf60010c