Set open file limit for dcmanager services
The current open file limit of 1024 is not enough when executing
operations that involve a large amount of subclouds in parallel
(e.g. patching orchestration). dcmanager-audit already has its limit
increased through the existing global option 'worker_rlimit_nofile',
but the other services are not using this option.

This commit moves the function that sets the limit to the utils.py
module and makes the other dcmanager services use the defined limit.

Test Plan:
1. PASS - Verify that the open file limit gets updated to the defined
          value by checking the service log file and by verifying the
          /proc/<pid>/limits file;
2. PASS - Verify that the new limit it enough to support DC patch
          orchestration with MAX_SUBCLOUD_GROUP_MAX_PARALLEL_SUBCLOUDS
          subclouds;
3. PASS - Verify that the limit can be changed by modifying the
          /etc/dcmanager/dcmanager.conf file with worker_rlimit_nofile
          option;
4. PASS - Verify that it's not possible to increase the soft limit past
          the hard limit.

Story: 2010584
Task: 47751

Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>
Change-Id: Ia52e532fd14cd2e61714477a561674d16ca45195