Ensure proper handling of special password
Currently subcloud prestage would fail if the sysadmin password
that starts with a special character (e.g. @Password123). This
is because the inventory file is generated with the password as-is
which results in yaml parsing failure by ansible.

This commit ensures the sysadmin password that starts with
a special character is handled properly.

Test Plan:
  - Verify successful subcloud add with password that starts with
    or without a special character.
  - Verify successful subcloud reinstall with password that starts
    with or without a special character.
  - Verify successful subcloud reconfigure with password that starts
    with or without a special character.
  - Verify successfull subcloud backup and restore with password
    that starts with or without a special character.
  - Verify successful subcloud prestage with password that starts
    with or without a special character.

Closes-Bug: 2015116
Signed-off-by: Tee Ngo <tee.ngo@windriver.com>
Change-Id: Ieedaee16ed0489c801cd56b1891daf71a26f1e74