update upgrade orch to not remove applied patches
In this commit, the DC upgrade orchestrator was updated to not remove
applied patches from the subcloud when the patch is not applied
on system controller.

Test Plan:
PASS: Upgrade system controller to 22.12, then apply a patch
on the subcloud and run the upgrade orchestration and verify
that the orchestration finished successfully.
PASS: Apply 21.12 patch on system controller, then commit the patch
and run patch orchestration. Then upgrade the system controller and
the subcloud and verify that the orchestration finished successfully.
PASS: Apply 21.12 patch on the system controller, then run patch
orchestration and upgrade the system controller and the subcloud
and verify that the orchestration finished successfully.
PASS: Upgrade system controller to 22.12, then apply a patch on
system controller and upgrade the subcloud and verify that
the orchestration reached the starting upgrade stage.

Closes-Bug: 2016701

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: Id914413c99a46782170300715b78d3129bd46720