Set subcloud to offline before reinstalling
In this commit, the DC upgrade orchestrator was updated to make
sure the subcloud would be offline during installation and migrating
data stages. This was changed to avoid misleading output when
checking the subcloud availability. Before this change, during the
mentioned stages, the availability would be online, even though it
was not possible to ping the subcloud. When the subcloud deploy
status changes to migrated, install failed or migration failed,
the audit will be triggered and the availability will be set
accordingly.

Test Plan:
PASS: upgrade system controller and sx subcloud and when orchestration
reach upgrade simplex, verify that subcloud is offline. Once subcloud
deploy status changes to migrated, verify that the subcloud is online.

Closes-Bug: 2017012

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: Ie13e5750fabd35cc91689d5b7f9d3e09f637bba7