Set sw version in db before install prep
During subcloud install, the process fails because the software
version is passed as an argument to the db update. This occurs because
the prep function from SubcloudInstall removes the key from the payload.
The software version update was changed to the existing db update before
the prep function call.

TEST PLAN:
- Run "dcmanager subcloud add" and see if the install proccess starts.
- Run "dcmanager subcloud reinstall" and see if the install process
  starts.
- Run "dcmanager subcloud restore --with-install" and see if the install
  process starts.

Story: 2010611
Task: 47709

Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>
Change-Id: I71c117b74f6e26f226c5af0c39f353dffc7ef0df