Send system local CA key to subcloud in upgrade
The distcloud component of this issue is related to the fact that as
of right now, for DX subclouds upgrade, the DC orchestrator calls
sysinv API to update the subcloud certificate stored in the
SystemController's 'system-local-ca' secret with the mode
CERT_MODE_OPENLDAP_CA. As the request does not include the key, the
sysinv conductor can only create an Opaque secret in the subcloud.

When migrate_platform_certificates_to_certmanager.yml playbook is
used, it stores both the 'system-local-ca' cert and key as a TLS
secret, which is used to issue platform certs for the subclouds.

Currently, during platform upgrade, the 'system-local-ca' in a DX
subcloud is always replaced by an Opaque secret containing only the
cert. If this secret was TLS (prior to the upgrade), it will be
impossible for cert-manager to reissue the subcloud's platform certs
due to not having the key stored anymore (the Opaque secret only has
the cert) after the upgrade.

The fix is to always send the key along with the cert to subcloud
during the orchestrated upgrade with CERT_MODE_OPENLDAP_CA mode.
Sysinv conductor will determine if the key should be stored (TLS) or
not (Opaque).

Test plan:
- Deploy SX, DX and DC with both SX and DX subclouds.
- Upgrade SX, DX and DC Systems with SX and DX subclouds from 21.12
  and 22.06 to designer iso 22.12.(*)(**)

P.S.:
(*) Due to the existence of an upgrade start script called in the
    'from' side that will overwrite the secret after this code is
    called, this change will only have effects in upgrades moving
    forward.
(**) The call for conductor API change was also tested by forcefully
     enabling the CLI API in a live system, and then installing a
     certificate through it with 'openldap_ca' mode (used by
     orchestration). This is not a formal test for this change, but
     is useful to verify it.

Partial-Bug: 2012435

Change-Id: Id14a049478ef69f7ce9cbfc9185b6a68bb8bd1cf
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>