Ignore certain alarms during kube upgrade pre-check
In this commit, the kubernetes upgrade orchestrator was updated to
not fail in the pre-check stage when certain alarms are set. Before
this change it was not possible to retry the kubernetes upgrade for
a subcloud that failed because of the 900.007 alarm.

Test Plan:
PASS: Run kubernetes upgrade orchestration and re-run
the orchestration for a subcloud that failed and verify that
the orchestration finished succesfully
PASS: Set a management-affecting alarm that is not in the following
list: ['100.003', '200.001', '700.004', '750.006','900.007', '900.401']
on the subcloud and verify that the kubernetes upgrade orchestration
failed for that subcloud.

Closes-Bug: 2012300

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: I25b97a98441db4bdbd1d3e546eda0b5a1588dcb1